[tests] Folder restructuring
